Extract cli/core; move vote/harvest/admin under issues
