+++ gittensor/cli/core/__init__.py
@@ -1,0 +1,4 @@
+# The MIT License (MIT)
+# Copyright © 2025 Entrius
+
+"""Shared CLI infrastructure reused by every command group."""
