+++ gittensor/cli/issue_commands/__init__.py
@@ -11,32 +11,21 @@ Command structure:
         register                     Register a new issue bounty
         bounty-pool                  View total bounty pool
         pending-harvest              View pending emissions
-    gitt harvest                 - Harvest emissions (top-level)
-    gitt vote                    - Validator consensus commands
-    gitt admin (alias: a)        - Owner-only commands
-        info                         View contract configuration
-        cancel-issue                 Cancel an issue
-        payout-issue                 Manual payout fallback
-        set-owner                    Transfer ownership
-        set-treasury                 Change treasury hotkey
+        harvest                      Harvest emissions
+        vote                         Validator consensus commands
+        admin (alias: a)             Owner-only commands
+            info                         View contract configuration
+            cancel-issue                 Cancel an issue
+            payout-issue                 Manual payout fallback
+            set-owner                    Transfer ownership
+            set-treasury                 Change treasury hotkey
 """
 
 import click
 
-from .admin import admin
-from .help import StyledGroup
+from gittensor.cli.core.help import StyledAliasGroup
 
-# Re-export helpers
-from .helpers import (
-    CONFIG_FILE,
-    GITTENSOR_DIR,
-    NETWORK_MAP,
-    console,
-    get_contract_address,
-    load_config,
-    read_issues_from_contract,
-    resolve_network,
-)
+from .admin import admin
 from .mutations import (
     issue_harvest,
     issue_register,
@@ -46,7 +35,7 @@ from .view import admin_info, issues_bounty_pool, issues_list, issues_pending_ha
 from .vote import vote
 
 
-@click.group(name='issues', cls=StyledGroup)
+@click.group(name='issues', cls=StyledAliasGroup)
 def issues_group():
     """Manage issue bounties, submissions, and predictions."""
     pass
@@ -57,6 +46,10 @@ issues_group.add_command(issues_submissions, name='submissions')
 issues_group.add_command(issue_register, name='register')
 issues_group.add_command(issues_bounty_pool, name='bounty-pool')
 issues_group.add_command(issues_pending_harvest, name='pending-harvest')
+issues_group.add_command(issue_harvest, name='harvest')
+issues_group.add_command(vote, name='vote')
+issues_group.add_command(admin)
+issues_group.add_alias('admin', 'a')
 
 # Add info to admin group
 admin.add_command(admin_info, name='info')
@@ -64,19 +57,8 @@ admin.add_command(admin_info, name='info')
 
 def register_commands(cli):
     """Register all issue-related commands with the root CLI group."""
-    # Issues group with alias
     cli.add_command(issues_group, name='issues')
     cli.add_alias('issues', 'i')
-
-    # Harvest as top-level command
-    cli.add_command(issue_harvest, name='harvest')
-
-    # Validator vote group
-    cli.add_command(vote, name='vote')
-
-    # Admin group with alias
-    cli.add_command(admin)
-    cli.add_alias('admin', 'a')
 
 
 __all__ = [
@@ -87,13 +69,4 @@ __all__ = [
     'issues_submissions',
     'issue_register',
     'issue_harvest',
-    # Helpers
-    'console',
-    'load_config',
-    'get_contract_address',
-    'resolve_network',
-    'read_issues_from_contract',
-    'GITTENSOR_DIR',
-    'CONFIG_FILE',
-    'NETWORK_MAP',
 ]
