+++ gittensor/cli/issue_commands/types.py
@@ -7,19 +7,9 @@ from __future__ import annotations
 
 import click
 
-from .helpers import MAX_ISSUE_ID, MAX_ISSUE_NUMBER, REPO_PATTERN, validate_ss58_address
+from gittensor.constants import MAX_ISSUE_ID
 
-
-class RepoNameType(click.ParamType):
-    """Owner/repo string matching ``REPO_PATTERN``."""
-
-    name = 'repo'
-
-    def convert(self, value, param, ctx):
-        trimmed = value.strip()
-        if not REPO_PATTERN.match(trimmed):
-            self.fail(f"'{value}' is not a valid owner/repo", param, ctx)
-        return trimmed
+from .helpers import MAX_ISSUE_NUMBER
 
 
 class ContractIssueType(click.IntRange):
@@ -40,21 +30,6 @@ class GitHubIssueType(click.IntRange):
         super().__init__(min=1, max=MAX_ISSUE_NUMBER)
 
 
-class SS58AddressType(click.ParamType):
-    """SS58 address validated via ``validate_ss58_address``."""
-
-    name = 'ss58'
-
-    def convert(self, value, param, ctx):
-        name = (param.name if param else None) or 'address'
-        try:
-            return validate_ss58_address(value, name)
-        except click.BadParameter as exc:
-            self.fail(str(exc), param, ctx)
-
-
 # Stateless singletons - one instance per type for the whole CLI
-REPO = RepoNameType()
 CONTRACT_ISSUE = ContractIssueType()
 GITHUB_ISSUE = GitHubIssueType()
-SS58 = SS58AddressType()
